ADDED: network client changes
