+++ PerseusMeteo/MeteoDataAPIs/OpenWeather/OpenWeatherAPI.swift
@@ -1,0 +1,120 @@
+//
+//  OpenWeatherAPI.swift
+//  TheDarkMoon
+//
+//  Created by Mikhail Zhigulin in 7531.
+//
+//  Copyright © 7531 - 7533 Mikhail A. Zhigulin of Novosibirsk
+//  Copyright © 7533 PerseusRealDeal
+//
+//  The year starts from the creation of the world according to a Slavic calendar.
+//  September, the 1st of Slavic year. For instance, "Sep 01, 2025" is the beginning of 7534.
+//
+//  See LICENSE for details. All rights reserved.
+//
+
+import Foundation
+
+public let weatherSchemeBase = "https://api.openweathermap.org/data/2.5/"
+public let weatherSchemeAttributes = "%@?lat=%@&lon=%@&appid=%@"
+
+public let geocodingDirectSchemeBase = "http://api.openweathermap.org/geo/1.0/"
+public let geocodingDirectSchemeAttributes = "direct?q=%@&limit=%@&appid=%@"
+
+public func prepareDirectURLString(cityName: String, limit: Int, appid: String) -> String {
+
+    let args: [String] = [cityName, "\(limit)", appid]
+    let attributes = String(format: geocodingDirectSchemeAttributes, arguments: args)
+
+    let urlString = geocodingDirectSchemeBase + attributes
+
+    return urlString
+}
+
+public enum OpenWeatherRequest: String {
+    case currentWeather = "weather" // Default.
+    case forecast = "forecast"
+}
+
+public enum Units: String {
+    case standard // Default.
+    case metric
+    case imperial
+}
+
+public enum Mode: String {
+    case json // Default.
+    case xml
+    case html
+}
+
+public struct Lang: RawRepresentable {
+    public var rawValue: String
+    public static let byDefault = Lang(rawValue: "")
+
+    public init(rawValue: String) {
+        self.rawValue = rawValue
+    }
+}
+
+extension Lang {
+    public static let en = Lang(rawValue: "en")
+    public static let ru = Lang(rawValue: "ru")
+}
+
+public struct OpenWeatherRequestData {
+
+    public let appid: String
+    public let request: OpenWeatherRequest
+
+    public let lat: String
+    public let lon: String
+
+    public let units: Units
+    public let lang: Lang
+    public let mode: Mode
+
+    // A number of timestamps, which will be returned in the API response.
+    public var cnt: Int = -1
+
+    public init(appid: String,
+                request: OpenWeatherRequest = .currentWeather,
+                lat: String = "55.66",
+                lon: String = "85.62",
+                units: Units = .standard,
+                lang: Lang = Lang.byDefault,
+                mode: Mode = Mode.json) {
+
+        self.appid = appid
+        self.request = request
+        self.lat = lat
+        self.lon = lon
+        self.units = units
+        self.lang = lang
+        self.mode = mode
+    }
+
+    public var urlString: String {
+
+        let args: [String] = [request.rawValue, lat, lon, appid]
+        var attributes = String(format: weatherSchemeAttributes, arguments: args)
+
+        if !lang.rawValue.isEmpty {
+            attributes.append("&lang=\(lang.rawValue)")
+        }
+
+        if request == .forecast && cnt != -1 {
+            attributes.append("&cnt=\(cnt)")
+        }
+
+        if mode != .json {
+            attributes.append("&mode=\(mode.rawValue)")
+        }
+
+        if units != .standard {
+            attributes.append("&units=\(units.rawValue)")
+        }
+
+        return weatherSchemeBase + attributes
+    }
+}
